+++ DashAI/front/src/tabs/Results.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { useParams } from "react-router-dom";
-import ExperimentsDrawer from "../components/ExperimentsDrawer";
-import RunsTable from "../components/RunsTable";
+import ExperimentsDrawer from "../components/results/ExperimentsDrawer";
+import RunsTable from "../components/results/RunsTable";
 /**
  * This component renders a table that shows the runs of the experiments and a drawer to select the experiment to visualize
  */
